+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+ - `memtx` MVCC memory monitoring
+
 ## [0.15.0] - 2022-08-09
 ### Fixed